requirements.txt
azure-functions
textblob 
vaderSentiment
numpy
requests
